+++ sdk/resourcemanager/maintenance/armmaintenance/zz_generated_configurations_client.go
@@ -39,7 +39,7 @@ func NewConfigurationsClient(subscriptionID string, credential azcore.TokenCrede
 	if options == nil {
 		options = &arm.ClientOptions{}
 	}
-	ep := cloud.AzurePublicCloud.Services[cloud.ResourceManager].Endpoint
+	ep := cloud.AzurePublic.Services[cloud.ResourceManager].Endpoint
 	if c, ok := options.Cloud.Services[cloud.ResourceManager]; ok {
 		ep = c.Endpoint
 	}
@@ -57,6 +57,7 @@ func NewConfigurationsClient(subscriptionID string, credential azcore.TokenCrede
 
 // CreateOrUpdate - Create or Update configuration record
 // If the operation fails it returns an *azcore.ResponseError type.
+// Generated from API version 2021-05-01
 // resourceGroupName - Resource Group Name
 // resourceName - Maintenance Configuration Name
 // configuration - The configuration
@@ -97,9 +98,9 @@ func (client *ConfigurationsClient) createOrUpdateCreateRequest(ctx context.Cont
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2021-09-01-preview")
+	reqQP.Set("api-version", "2021-05-01")
 	req.Raw().URL.RawQuery = reqQP.Encode()
-	req.Raw().Header.Set("Accept", "application/json")
+	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, runtime.MarshalAsJSON(req, configuration)
 }
 
@@ -114,6 +115,7 @@ func (client *ConfigurationsClient) createOrUpdateHandleResponse(resp *http.Resp
 
 // Delete - Delete Configuration record
 // If the operation fails it returns an *azcore.ResponseError type.
+// Generated from API version 2021-05-01
 // resourceGroupName - Resource Group Name
 // resourceName - Maintenance Configuration Name
 // options - ConfigurationsClientDeleteOptions contains the optional parameters for the ConfigurationsClient.Delete method.
@@ -152,9 +154,9 @@ func (client *ConfigurationsClient) deleteCreateRequest(ctx context.Context, res
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2021-09-01-preview")
+	reqQP.Set("api-version", "2021-05-01")
 	req.Raw().URL.RawQuery = reqQP.Encode()
-	req.Raw().Header.Set("Accept", "application/json")
+	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil
 }
 
@@ -169,6 +171,7 @@ func (client *ConfigurationsClient) deleteHandleResponse(resp *http.Response) (C
 
 // Get - Get Configuration record
 // If the operation fails it returns an *azcore.ResponseError type.
+// Generated from API version 2021-05-01
 // resourceGroupName - Resource Group Name
 // resourceName - Maintenance Configuration Name
 // options - ConfigurationsClientGetOptions contains the optional parameters for the ConfigurationsClient.Get method.
@@ -207,9 +210,9 @@ func (client *ConfigurationsClient) getCreateRequest(ctx context.Context, resour
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2021-09-01-preview")
+	reqQP.Set("api-version", "2021-05-01")
 	req.Raw().URL.RawQuery = reqQP.Encode()
-	req.Raw().Header.Set("Accept", "application/json")
+	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil
 }
 
@@ -224,9 +227,10 @@ func (client *ConfigurationsClient) getHandleResponse(resp *http.Response) (Conf
 
 // NewListPager - Get Configuration records within a subscription
 // If the operation fails it returns an *azcore.ResponseError type.
+// Generated from API version 2021-05-01
 // options - ConfigurationsClientListOptions contains the optional parameters for the ConfigurationsClient.List method.
 func (client *ConfigurationsClient) NewListPager(options *ConfigurationsClientListOptions) *runtime.Pager[ConfigurationsClientListResponse] {
-	return runtime.NewPager(runtime.PageProcessor[ConfigurationsClientListResponse]{
+	return runtime.NewPager(runtime.PagingHandler[ConfigurationsClientListResponse]{
 		More: func(page ConfigurationsClientListResponse) bool {
 			return false
 		},
@@ -259,9 +263,9 @@ func (client *ConfigurationsClient) listCreateRequest(ctx context.Context, optio
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2021-09-01-preview")
+	reqQP.Set("api-version", "2021-05-01")
 	req.Raw().URL.RawQuery = reqQP.Encode()
-	req.Raw().Header.Set("Accept", "application/json")
+	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil
 }
 
@@ -276,6 +280,7 @@ func (client *ConfigurationsClient) listHandleResponse(resp *http.Response) (Con
 
 // Update - Patch configuration record
 // If the operation fails it returns an *azcore.ResponseError type.
+// Generated from API version 2021-05-01
 // resourceGroupName - Resource Group Name
 // resourceName - Maintenance Configuration Name
 // configuration - The configuration
@@ -315,9 +320,9 @@ func (client *ConfigurationsClient) updateCreateRequest(ctx context.Context, res
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2021-09-01-preview")
+	reqQP.Set("api-version", "2021-05-01")
 	req.Raw().URL.RawQuery = reqQP.Encode()
-	req.Raw().Header.Set("Accept", "application/json")
+	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, runtime.MarshalAsJSON(req, configuration)
 }