+++ sdk/resourcemanager/maintenance/armmaintenance/zz_generated_configurationsforresourcegroup_client.go
@@ -39,7 +39,7 @@ func NewConfigurationsForResourceGroupClient(subscriptionID string, credential a
 	if options == nil {
 		options = &arm.ClientOptions{}
 	}
-	ep := cloud.AzurePublicCloud.Services[cloud.ResourceManager].Endpoint
+	ep := cloud.AzurePublic.Services[cloud.ResourceManager].Endpoint
 	if c, ok := options.Cloud.Services[cloud.ResourceManager]; ok {
 		ep = c.Endpoint
 	}
@@ -57,11 +57,12 @@ func NewConfigurationsForResourceGroupClient(subscriptionID string, credential a
 
 // NewListPager - Get Configuration records within a subscription and resource group
 // If the operation fails it returns an *azcore.ResponseError type.
+// Generated from API version 2021-05-01
 // resourceGroupName - Resource Group Name
 // options - ConfigurationsForResourceGroupClientListOptions contains the optional parameters for the ConfigurationsForResourceGroupClient.List
 // method.
 func (client *ConfigurationsForResourceGroupClient) NewListPager(resourceGroupName string, options *ConfigurationsForResourceGroupClientListOptions) *runtime.Pager[ConfigurationsForResourceGroupClientListResponse] {
-	return runtime.NewPager(runtime.PageProcessor[ConfigurationsForResourceGroupClientListResponse]{
+	return runtime.NewPager(runtime.PagingHandler[ConfigurationsForResourceGroupClientListResponse]{
 		More: func(page ConfigurationsForResourceGroupClientListResponse) bool {
 			return false
 		},
@@ -98,9 +99,9 @@ func (client *ConfigurationsForResourceGroupClient) listCreateRequest(ctx contex
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2021-09-01-preview")
+	reqQP.Set("api-version", "2021-05-01")
 	req.Raw().URL.RawQuery = reqQP.Encode()
-	req.Raw().Header.Set("Accept", "application/json")
+	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil
 }